# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "context_available": {
            "title": "Context Available",
            "type": "boolean"
          },
          "eligibility": {
            "title": "Eligibility",
            "type": "string"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "in_selection": {
            "title": "In Selection",
            "type": "integer"
          },
          "is_free": {
            "title": "Is Free",
            "type": "boolean"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "recommendation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Reason"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "seeders": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seeders"
          },
          "site_count": {
            "title": "Site Count",
            "type": "integer"
          },
          "site_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Name"
          },
          "site_options": {
            "items": {
              "$ref": "#/components/schemas/SiteOption"
            },
            "title": "Site Options",
            "type": "array"
          },
          "site_selection_reason": {
            "title": "Site Selection Reason",
            "type": "string"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "volume_factor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Volume Factor"
          }
        },
        "required": [
          "id",
          "title",
          "site_name",
          "size",
          "seeders",
          "resolution",
          "codec",
          "group_name",
          "recommendation",
          "recommendation_reason",
          "eligibility",
          "in_selection",
          "context_available",
          "is_free",
          "volume_factor",
          "site_count",
          "site_options",
          "site_selection_reason"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "Film": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "imdb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Id"
          },
          "issue_labels": {
            "items": {
              "type": "string"
            },
            "title": "Issue Labels",
            "type": "array"
          },
          "issues": {
            "items": {
              "enum": [
                "no_eligible",
                "submit_failed",
                "context_expired"
              ],
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          },
          "library_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library State"
          },
          "original_title": {
            "title": "Original Title",
            "type": "string"
          },
          "playlist_id": {
            "title": "Playlist Id",
            "type": "integer"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "rank_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank No"
          },
          "status": {
            "enum": [
              "unrecognized",
              "unchecked",
              "missing",
              "searching",
              "candidates",
              "selected",
              "downloading",
              "in_library"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "title": "Status Label",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "transfer": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "waiting_library",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "playlist_id",
          "rank_no",
          "title",
          "original_title",
          "year",
          "tmdb_id",
          "imdb_id",
          "library_state",
          "status",
          "status_label",
          "issues",
          "issue_labels",
          "transfer",
          "poster_url"
        ],
        "title": "Film",
        "type": "object"
      },
      "FilmDetail": {
        "properties": {
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "excluded_count": {
            "title": "Excluded Count",
            "type": "integer"
          },
          "excluded_summary": {
            "items": {
              "$ref": "#/components/schemas/ExcludedReason"
            },
            "title": "Excluded Summary",
            "type": "array"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/FilmHistoryRecord"
            },
            "title": "History",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "imdb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Id"
          },
          "issue_labels": {
            "items": {
              "type": "string"
            },
            "title": "Issue Labels",
            "type": "array"
          },
          "issues": {
            "items": {
              "enum": [
                "no_eligible",
                "submit_failed",
                "context_expired"
              ],
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          },
          "library_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library Checked At"
          },
          "library_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library State"
          },
          "original_title": {
            "title": "Original Title",
            "type": "string"
          },
          "playlist_id": {
            "title": "Playlist Id",
            "type": "integer"
          },
          "playlist_name": {
            "title": "Playlist Name",
            "type": "string"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "rank_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank No"
          },
          "search": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilmSearchSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "search_site_count": {
            "title": "Search Site Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "unrecognized",
              "unchecked",
              "missing",
              "searching",
              "candidates",
              "selected",
              "downloading",
              "in_library"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "title": "Status Label",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "transfer": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "waiting_library",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "playlist_id",
          "rank_no",
          "title",
          "original_title",
          "year",
          "tmdb_id",
          "imdb_id",
          "library_state",
          "status",
          "status_label",
          "issues",
          "issue_labels",
          "transfer",
          "poster_url",
          "backdrop_url",
          "playlist_name",
          "library_checked_at",
          "candidates",
          "excluded_summary",
          "excluded_count",
          "history",
          "search",
          "search_site_count"
        ],
        "title": "FilmDetail",
        "type": "object"
      },
...
      "FilmPage": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Film"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistSummary"
            },
            "title": "Playlists",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "counts",
          "playlists"
        ],
        "title": "FilmPage",
        "type": "object"
      },
      "FilmSearchStarted": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "total"
        ],
        "title": "FilmSearchStarted",
        "type": "object"
      },
...
      "FilmTmdbPayload": {
        "description": "Manually pin a film to a TMDB movie id when automatic recognition picked the wrong one.",
        "properties": {
          "tmdb_id": {
            "maximum": 100000000.0,
            "minimum": 1.0,
            "title": "Tmdb Id",
            "type": "integer"
          }
        },
        "required": [
          "tmdb_id"
        ],
        "title": "FilmTmdbPayload",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaylistSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          }
        },
        "required": [
          "id",
          "name",
          "source_type",
          "item_count"
        ],
        "title": "PlaylistSummary",
        "type": "object"
      },
...
      "TmdbMatch": {
        "properties": {
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "original_title": {
            "title": "Original Title",
            "type": "string"
          },
          "overview": {
            "title": "Overview",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "title": "Tmdb Id",
            "type": "integer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "tmdb_id",
          "title",
          "original_title",
          "year",
          "overview",
          "current"
        ],
        "title": "TmdbMatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AutoList"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/films": {
      "get": {
        "operationId": "films_api_films_get",
        "parameters": [
          {
            "in": "query",
            "name": "playlist_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Playlist Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 120,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 60,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Films"
      }
    },
    "/api/films/{item_id}": {
      "get": {
        "operationId": "film_detail_api_films__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Film Detail"
      }
    },
    "/api/films/{item_id}/recognize": {
      "post": {
        "description": "按导入时的原名、年份与 IMDb 重新识别这一部影片（整体替换原有识别结果）。",
        "operationId": "recognize_film_api_films__item_id__recognize_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recognize Film"
      }
    },
    "/api/films/{item_id}/search": {
      "post": {
        "operationId": "search_film_api_films__item_id__search_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmSearchStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Film"
      }
    },
    "/api/films/{item_id}/tmdb": {
      "post": {
        "description": "手动指定 TMDB 影片，覆盖自动识别结果。",
        "operationId": "set_film_tmdb_api_films__item_id__tmdb_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmTmdbPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Film Tmdb"
      }
    },
    "/api/films/{item_id}/tmdb-matches": {
      "get": {
        "description": "手动指定用：按关键词在 TMDB 搜索候选影片。",
        "operationId": "tmdb_matches_api_films__item_id__tmdb_matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 120,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TmdbMatch"
                  },
                  "title": "Response Tmdb Matches Api Films  Item Id  Tmdb Matches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tmdb Matches"
      }
    },
...
  }
}